+++ src/Utilities/Config.php
@@ -39,15 +39,14 @@ class Config
         OPTIONS_UPDATE_FAILED                            = 'jtlconnector_update_failed',
         OPTIONS_USE_DELIVERYTIME_CALC                    = 'jtlconnector_use_deliverytime_calc',
         OPTIONS_PRAEFIX_DELIVERYTIME                     = 'jtlconnector_praefix_deliverytime',
-        OPTIONS_DEVELOPER_LOGGING                        = 'developer_logging',
+        OPTIONS_DEVELOPER_LOGGING                        = ConfigSchema::DEBUG,
         OPTIONS_AUTO_WOOCOMMERCE_OPTIONS                 = 'jtlconnector_auto_woocommerce',
         OPTIONS_AUTO_GERMAN_MARKET_OPTIONS               = 'jtlconnector_auto_german_market',
         OPTIONS_CONSIDER_SUPPLIER_INFLOW_DATE            = 'jtlconnector_consider_supplier_inflow_date',
         OPTIONS_CUSTOM_CHECKOUT_FIELDS                   = 'jtlconnector_custom_checkout_fields',
         OPTIONS_LIMIT_CUSTOMER_QUERY_TYPE                = 'jtlconnector_limit_customer_query',
         OPTIONS_FEATURES_JSON                            = 'jtlconnector_features_json',
-        OPTIONS_IGNORE_ORDERS_YOUNGER_THAN               = 'jtlconnector_ignore_orders_younger_than',
-        OPTIONS_USE_CACHE                                = 'jtlconnector_disable_cache';
+        OPTIONS_IGNORE_ORDERS_YOUNGER_THAN               = 'jtlconnector_ignore_orders_younger_than';
 
     public const JTLWCC_CONFIG_DEFAULTS = [
         //FIRSTPAGE
@@ -87,7 +86,6 @@ class Config
         Config::OPTIONS_AUTO_WOOCOMMERCE_OPTIONS => true,
         Config::OPTIONS_AUTO_GERMAN_MARKET_OPTIONS => true,
         Config::OPTIONS_AUTO_B2B_MARKET_OPTIONS => true,
-        Config::OPTIONS_USE_CACHE => true
     ];
 
     public const JTLWCC_CONFIG = [
@@ -117,7 +115,6 @@ class Config
         Config::OPTIONS_IGNORE_ORDERS_YOUNGER_THAN => 'int',
         //Page
         Config::OPTIONS_DEVELOPER_LOGGING => 'bool',
-        Config::OPTIONS_USE_CACHE => 'bool',
         Config::OPTIONS_AUTO_WOOCOMMERCE_OPTIONS => 'bool',
         Config::OPTIONS_AUTO_GERMAN_MARKET_OPTIONS => 'bool',
         Config::OPTIONS_AUTO_B2B_MARKET_OPTIONS => 'bool',